DjangoBackend/requirements-dev.txt
asgiref==3.2.3
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.2
django-cors-headers==3.2.1
djangorestframework==3.11.0
python-decouple==3.3
pytz==2019.3
sqlparse==0.3.0
static3==0.7.0
